moved project files from Filehandler
