--- ViAn/Filehandler/filesysmanager.cpp
@@ -1,5 +1,0 @@
-#include "filesysmanager.h"
-
-FilesysManager::FilesysManager()
-{
-}
