--- ViAn/Filehandler/filesysmanager.h
@@ -1,10 +1,0 @@
-#ifndef FILESYSMANAGER_H
-#define FILESYSMANAGER_H
-
-class FilesysManager
-{
-public:
-    FilesysManager();
-};
-
-#endif // FILESYSMANAGER_H
