+++ ViAn/Filehandler/saveable.h
@@ -12,10 +12,10 @@ class Saveable
 public:
     std::string save_name;
     Saveable();
+
     virtual ~Saveable();
     virtual void read(const QJsonObject& json) = 0;
     virtual void write(QJsonObject& json) = 0;
-
 };
 
 #endif // SAVABLE_H
